Add --no-reconnect option to disable reconnect timer
When idle is available, the main purpose of the timer loop is to reconnect the
socket when it is disconnected.

An alternative is to allow systemd to handle this for us. We can bind the daemon
to the mpd service, so that it is started automatically with the mpd service.

Add a new command-line option `--no-reconnect`. If this is set, the daemon will
exit cleanly instead of attempting a reconnect loop, under the assumption that
systemd (or any other service manager) will restart it when mpd is available
again.

Remove the WantedBy=default.target so the daemon is only started with mpd.